group filter hotfix

+++ app/controllers/index_controller.php
@@ -13,9 +13,9 @@ class IndexController extends Page {
 
     public function indexAction() {
         $conditions = null;
-        $this->params['group'] = (integer) ($this->params['group'] ?? 0);
+        $this->params['group'] = $this->params['group'] ?? 0;
         if(!empty($this->params['group'])):
-            $conditions = $this->params['group'];
+            $conditions = (integer) $this->params['group'];
         endif;
         $this->projects = $this->Project->Find($conditions);
         $this->groups = $this->ProjectGroup->Find();
